+++ src/main/java/uz/backall/sell/sellAmountDate/SellAmountDateService.java
@@ -148,7 +148,7 @@ public class SellAmountDateService {
       .orElseThrow(() -> new SellAmountDateNotFoundException("No SellAmountDate found for storeId: " + storeId));
   }
 
-  public Double getMonthAmount(Long storeId) {
+  public Long getMonthAmount(Long storeId) {
     Calendar calendar = Calendar.getInstance();
     int monthNumber = calendar.get(Calendar.MONTH) + 1; // Months are 0-based in Calendar
     int year = calendar.get(Calendar.YEAR);
@@ -156,7 +156,7 @@ public class SellAmountDateService {
     // Construct the date pattern for the given month and year (e.g., "2024-08-%")
     String datePattern = year + "-" + (monthNumber < 10 ? "0" : "") + monthNumber + "-%";
 
-    Double totalAmount = sellAmountDateRepository.findTotalAmountByStoreIdAndDatePattern(storeId, datePattern);
-    return totalAmount != null ? totalAmount : 0.0;
+    Long totalAmount = sellAmountDateRepository.findTotalAmountByStoreIdAndDatePattern(storeId, datePattern);
+    return totalAmount != null ? totalAmount : 0;
   }
 }
